Major update. See desc
So, to begin with scripts where updated in order to compile the
component into lib folder with keeping all versions of the files. Flow
support works fine now with compiled files. Removed rollupjs, as we
compile the component with babeljs. Applied prettier on package.json to
improve readability and formatting.